restructured app pipeline code into flow package
