+++ internal/app/flow/pipe.go
@@ -1,0 +1,7 @@
+package flow
+
+import "github.com/urfave/cli/v2"
+
+type Pipe interface {
+	Run(ctx *cli.Context) error
+}
